fix(driver): require daemon-backed calls and stabilize E2E (#2338)
* test(driver): run element-token protocol through daemon

* test(driver): model capture scope protocol gates

* test(driver): record compositor canaries after setup

* test(driver): drain sentinel setup input

* docs(driver): regenerate daemon-backed references

* test(driver): keep compositor sentinel responsive

* test(driver): assert daemon policy error envelope

* test(driver): drive compositor heartbeat over IPC

* test(driver): probe Windows daemon over named pipe

* fix(driver): wait for real DevTools endpoint

* fix(driver): stabilize daemon-backed E2E contracts

* test(driver): isolate compositor heartbeat loop

* fix(driver): repair remaining hosted E2E regressions

* test(driver): use supported compositor leak canary

* fix(driver): restore Chromium approval endpoint

* fix(driver): target nested compositor windows by pid

* fix(driver): use explicit pid for compositor input

* fix(driver): target compositor process families

* fix(driver): bound compositor recording snapshots

* fix(driver): avoid compositor recording re-resolution

* fix(driver): preserve focused compositor keyboard delivery

* fix(driver): avoid redundant compositor focus request

* test(driver): require renderer heartbeat in compositor E2E

* fix(driver): map compositor sentinel before page load

* fix(driver): avoid headless activation configure

* fix(driver): configure compositor activation once

* fix(driver): pair initial compositor activation

* fix(driver): defer headless seat focus until activation

* fix(driver): attach virtual compositor keyboard

* fix(driver): restore compositor state evidence

* fix(driver): complete Wayland browser routing

* fix(driver): complete Chromium Wayland protocols

* fix(driver): avoid double-counting Chromium document offsets

* fix(driver): focus desktop compositor pointer events

* fix(driver): preserve background child-window focus

* fix(driver): declare compositor toplevel type

* fix(driver): declare compositor focus helper

* fix(driver): fall back for inert Edge consent action

* test(driver): retry transient sentinel startup

---------

+++ libs/cua-driver/rust/crates/platform-linux/src/tools/page.rs
@@ -24,9 +24,9 @@ impl Default for LinuxPageBackend {
 
 #[async_trait]
 impl PageBackend for LinuxPageBackend {
-    async fn get_text(&self, pid: i32, window_id: u32) -> anyhow::Result<String> {
+    async fn get_text(&self, pid: i32, window_id: u64) -> anyhow::Result<String> {
         let pid_u = pid as u32;
-        let xid = window_id as u64;
+        let xid = window_id;
         let result = tokio::task::spawn_blocking(move || crate::atspi::walk_tree(pid_u, xid, None))
             .await
             .map_err(|e| anyhow::anyhow!("AT-SPI walk task failed: {e}"))?;
@@ -36,12 +36,12 @@ impl PageBackend for LinuxPageBackend {
     async fn query_dom(
         &self,
         pid: i32,
-        window_id: u32,
+        window_id: u64,
         css_selector: &str,
         _attributes: &[String],
     ) -> anyhow::Result<String> {
         let pid_u = pid as u32;
-        let xid = window_id as u64;
+        let xid = window_id;
         let selector = css_selector.to_owned();
         let result = tokio::task::spawn_blocking(move || crate::atspi::walk_tree(pid_u, xid, None))
             .await
@@ -105,7 +105,7 @@ impl PageBackend for LinuxPageBackend {
     async fn execute_javascript(
         &self,
         _pid: i32,
-        _window_id: u32,
+        _window_id: u64,
         javascript: &str,
     ) -> anyhow::Result<String> {
         let port: u16 = match std::env::var("CUA_DRIVER_CDP_PORT")
@@ -127,7 +127,7 @@ impl PageBackend for LinuxPageBackend {
     async fn execute_javascript_targeted(
         &self,
         pid: i32,
-        window_id: u32,
+        window_id: u64,
         javascript: &str,
         cdp_port: Option<u16>,
         target_url_contains: Option<&str>,
